% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate}
\usage{
calibrate(
  parameters,
  target,
  target_tt,
  summary_function,
  tolerance,
  interval = c(0.01, 2000)/365,
  ...
)
}
\arguments{
\item{parameters}{Other malariasimulation model parameters}

\item{target}{Values of target variable to calibrate to.}

\item{target_tt}{Timesteps of target.}

\item{summary_function}{A function that take the raw model output as an argument and
produces a vector of the target variable.}

\item{tolerance}{The routine will complete when the difference between the target variable
and each of the target values are all within this range. Tolerance is specified in
the units of the target variable (e.g. if my target variable is prevalence, and my tolerance is 0.05,
then the difference between the target variable and each of the target values must be <5\% for the routine to succeed).}

\item{interval}{The search interval of init EIRs.}

\item{...}{Additional arguments to pass to the \code{uniroot()} function.}
}
\value{
Uniroot output
}
\description{
General calibration function for malariasimulation. Please see the documentation for
\link[stats]{uniroot} to become familiar with possible additional arguments and convergence
errors.
}
