% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{check_elimination}
\alias{check_elimination}
\title{Elimination check}
\usage{
check_elimination(output, target)
}
\arguments{
\item{output}{Model run output}

\item{target}{Target}
}
\value{
Boolean indicator of unwanted elimination
}
\description{
Elimination check
}
