% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{check_fit}
\alias{check_fit}
\title{Check the fit with proposed EIR}
\usage{
check_fit(proposed_eir, parameters, target, summary_function)
}
\arguments{
\item{proposed_eir}{Proposed new EIR value}

\item{parameters}{Malariasimulation model parameter list}

\item{target}{Values of target variable to calibrate to.}

\item{summary_function}{A function that take the raw model output as an}
}
\value{
Boolean indicator of move success
}
\description{
Check the fit with proposed EIR
}
