% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcali.R
\name{default_change}
\alias{default_change}
\title{Default change function for standard parameters}
\usage{
default_change(x, parameters, par_name)
}
\arguments{
\item{x}{}

\item{parameters}{}

\item{par_name}{}
}
\value{

}
\description{
Default change function for standard parameters
}
