% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcali.R
\name{eir_change}
\alias{eir_change}
\title{Default change function that varies EIR}
\usage{
eir_change(x, parameters, par_name)
}
\arguments{
\item{x}{}

\item{parameters}{}

\item{par_name}{}
}
\value{

}
\description{
Default change function that varies EIR
}
