% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium.R
\name{eq_objective}
\alias{eq_objective}
\title{Equilibrium Objective function}
\usage{
eq_objective(eir, target_pfpr, ft)
}
\arguments{
\item{eir}{EIR for adults, in units of infectious bites per person per year}

\item{target_pfpr}{Target PfPr in 2-10 year olds}

\item{ft}{The proportion of clinical cases effectively treated}
}
\value{
The difference between current and target prevalence
}
\description{
Used to estimate the starting EIR for a given PfPr
}
