% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcali.R
\name{fit_spline}
\alias{fit_spline}
\title{Fit non-linear spline model to simulation data}
\usage{
fit_spline(
  sim_data,
  target,
  par_name,
  summary_name,
  trans.summary,
  trans.par,
  inv.trans.summary,
  inv.trans.par
)
}
\arguments{
\item{target}{Values of target variable to calibrate to.}

\item{sims}{data.table of simulation output generated by run_simulations}
}
\value{
A list of a data.table of predicted EIR values for a range of prevalences, predicted EIR values for specific target values, the spline model fit
}
\description{
Fit non-linear spline model to simulation data
}
