% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal.R
\name{from_real}
\alias{from_real}
\title{Transform unbounded to bounded}
\usage{
from_real(x, limits)
}
\arguments{
\item{x}{Value}

\item{limits}{Vector of lower and upper limits}
}
\value{
Transformed value
}
\description{
Transform unbounded to bounded
}
