% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium.R
\name{get_eq_eir}
\alias{get_eq_eir}
\title{Estimate the EIR for a target PfPr}
\usage{
get_eq_eir(target_pfpr, ft = 0.1)
}
\arguments{
\item{target_pfpr}{Target PfPr in 2-10 year olds}

\item{ft}{The proportion of clinical cases effectively treated}
}
\value{
EIR
}
\description{
Estimate the equilibrium EIR that results in a specified Plasmodium falciparum
prevalence in 2-10 year olds for a given treatment coverage
}
\examples{
get_eq_eir(target_pfpr = 0.2, ft = 0.4)
}
