% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{good_move}
\alias{good_move}
\title{Check if the move has been successful}
\usage{
good_move(objective, direction)
}
\arguments{
\item{objective}{Vector of objectives}

\item{direction}{Direction}
}
\value{
Boolean indicator of move success
}
\description{
Check if the move has been successful
}
