% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{linear_interpolate}
\alias{linear_interpolate}
\title{Interpolate EIR linearly}
\usage{
linear_interpolate(eir, objective)
}
\arguments{
\item{eir}{Vector of EIRs}

\item{objective}{Vector of objectives}
}
\value{
Interpolated EIR
}
\description{
Interpolate EIR linearly
}
