% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcali.R
\name{nlcali}
\alias{nlcali}
\title{Calibrate using non-linear model}
\usage{
nlcali(
  parameters,
  target,
  target_tt,
  summary_function,
  ncores = 1,
  nsims = 100
)
}
\arguments{
\item{parameters}{Other malariasimulation model parameters}

\item{target}{Values of target variable to calibrate to.}

\item{target_tt}{Timesteps of target.}

\item{summary_function}{A function that produces a vector of the target variable.}

\item{ncores}{Number of cpu cores to perform simulations on in parallel}

\item{nsims}{Number of simulations to run}
}
\value{
A list of data from simulation runs, predicted EIR value for calibration targets, the fit of the non-linear model object
}
\description{
Calibrate using non-linear model
}
