% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{Objective function}
\usage{
objective(
  x,
  parameters,
  summary_function,
  target,
  weights,
  tolerance,
  elimination_check
)
}
\arguments{
\item{x}{Input parameter to be varied. For malariasimulation this is EIR.}

\item{parameters}{Other malariasimulation model parameters}

\item{summary_function}{A function that take the raw model output as an argument and
produces a vector of the target variable.}

\item{target}{Values of target variable to calibrate to.}

\item{weights}{A numerical vector of weights the same length as target giving the weights to use for elements of target.}

\item{tolerance}{The routine will complete when the average absolute weighted difference between the target variable
and the target values falls below this value. Tolerance is specified in
the units of the target variable (e.g. if my target variable is prevalence, and my tolerance is 0.05,
then the average absolute sum of the weighted difference between the target variable and the target values must be <5\% for the routine to succeed).}

\item{elimination_check}{Check transmission is maintained for all target points with ongoing transmission before exiting early.}
}
\value{
Difference between output and target.
}
\description{
Quantifies distance to target
}
