% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium.R
\name{pfpr_2_10}
\alias{pfpr_2_10}
\title{Equilibrium PfPr}
\usage{
pfpr_2_10(eir, ft = 0.1)
}
\arguments{
\item{eir}{EIR for adults, in units of infectious bites per person per year}

\item{ft}{The proportion of clinical cases effectively treated}
}
\value{
Plamodium falciparum prevalence in 2-10 year olds
}
\description{
Estimate equilibrium pfpr prevalence for a given eir and treatment coverage
}
