% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{print_update}
\alias{print_update}
\title{Update message}
\usage{
print_update(eir, objective)
}
\arguments{
\item{eir}{Vector of EIRs}

\item{objective}{Vector of objectives}
}
\description{
Update message
}
