% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal.R
\name{proposal}
\alias{proposal}
\title{Propose a new value for EIR}
\usage{
proposal(current_eir, limits, direction, step = 0.5)
}
\arguments{
\item{current_eir}{Current value}

\item{limits}{Inferred limits}

\item{direction}{"decrease" or "increase"}

\item{step}{The step size}
}
\value{
New EIR value
}
\description{
Propose a new value for EIR
}
