% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcali.R
\name{run_simulations}
\alias{run_simulations}
\title{Run simulations from malariasimuation}
\usage{
run_simulations(
  parameters,
  target,
  target_tt,
  test_EIRs,
  ncores,
  summary_function
)
}
\arguments{
\item{parameters}{Other malariasimulation model parameters}

\item{target}{Values of target variable to calibrate to.}

\item{target_tt}{Timesteps of target.}

\item{test_EIRs}{A vector of EIR values to use in the simulations}

\item{ncores}{Number of simulations to run}

\item{summary_function}{A function that produces a vector of the target variable.}
}
\value{
A data.table of simulation results
}
\description{
Run simulations from malariasimuation
}
