% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcali.R
\name{score_calibration}
\alias{score_calibration}
\title{Simulate from calibrated EIR value and calculate continuous ranked probability score}
\usage{
score_calibration(
  out_EIR,
  parameters,
  target,
  target_tt,
  ncores = 1,
  nsims = 100,
  summary_function
)
}
\arguments{
\item{out_EIR}{fitted EIR from a cali or ncali run}

\item{parameters}{Other malariasimulation model parameters}

\item{target}{Values of target variable to calibrate to.}

\item{target_tt}{Timesteps of target.}

\item{ncores}{Number of cpu cores to perform simulations on in parallel}

\item{nsims}{Number of simulations to run}

\item{summary_function}{A function that produces a vector of the target variable.}
}
\value{

}
\description{
Simulate from calibrated EIR value and calculate continuous ranked probability score
}
