% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_helpers.R
\name{summary_pfpr_2_10}
\alias{summary_pfpr_2_10}
\title{Prevalence 2-10 year olds}
\usage{
summary_pfpr_2_10(x)
}
\arguments{
\item{x}{malariasimualtion model output}
}
\value{
A vector of PfPR 2-10.
}
\description{
Prevalence 2-10 year olds
}
