% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal.R
\name{to_real}
\alias{to_real}
\title{Transform bounded to unbounded}
\usage{
to_real(x, limits)
}
\arguments{
\item{x}{Value}

\item{limits}{Vector of lower and upper limits}
}
\value{
Transformed value
}
\description{
Transform bounded to unbounded
}
