% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate}
\usage{
calibrate(
  parameters,
  target,
  summary_function,
  eq_prevalence,
  eq_ft = 0,
  human_population = c(1000, 10000, 1e+05),
  eir_limits = c(0, 1500),
  max_attempts = 10
)
}
\arguments{
\item{parameters}{Malariasimulation model parameter list}

\item{target}{Values of target variable to calibrate to.}

\item{summary_function}{A function that take the raw model output as an}

\item{eq_prevalence}{Estimate of baseline prevalence (to inform EIR initialisation)}

\item{eq_ft}{Estimate of baseline treatment coverage (to inform EIR initialisation)}

\item{human_population}{Vector of increasing population sizes.}

\item{eir_limits}{Limits of EIR}

\item{max_attempts}{Maximum number of calibration attempts (model runs)}
}
\value{
EIR estimate
}
\description{
General calibration function for malariasimulation.
}
