% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate}
\usage{
calibrate(
  target,
  summary_function,
  parameters,
  tolerance,
  weights = rep(1, length(target)),
  elimination_check = TRUE,
  maxiter = 20,
  low = 0.001,
  high = 2000
)
}
\arguments{
\item{target}{Values of target variable to calibrate to.}

\item{summary_function}{A function that take the raw model output as an argument and
produces a vector of the target variable.}

\item{parameters}{Other malariasimulation model parameters}

\item{tolerance}{The routine will complete when the absolute sum of the weighted difference between the target variable
and the target values falls below this value. Tolerance is specified in
the units of the target variable (e.g. if my target variable is prevalence, and my tolerance is 0.05,
then the absolute sum of the weighted difference between the target variable and the target values must be <5\% for the routine to succeed).}

\item{weights}{A numerical vector of weights the same length as target giving the weights to use for elements of target.}

\item{elimination_check}{Check transmission is maintained for all target points with ongoing transmission before exiting early.}

\item{maxiter}{Maximum iterations. If stopping criteria have not been met by \code{maxiter}, the current best estimate will be returned}

\item{low}{Lower boundof EIRs}

\item{high}{Upper bound of EIRs}
}
\value{
Uniroot output
}
\description{
General calibration function for malariasimulation.
}
