% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate}
\usage{
calibrate(
  parameters,
  target,
  target_tt,
  summary_function,
  tolerance,
  interval = c(0.01, 2000)/365,
  ...
)
}
\arguments{
\item{parameters}{Other malariasimulation model parameters}

\item{target}{Values of target variable to calibrate to.}

\item{target_tt}{Timesteps of target.}

\item{summary_function}{A function that produces a vector of the target variable.}

\item{tolerance}{The average tolerance, in target variable units, with which to match target.}

\item{interval}{The search interval.}

\item{...}{Additional arguments to pass to the \code{uniroot()} function.}
}
\value{
Uniroot output
}
\description{
Calibrate
}
