% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{Objective function}
\usage{
objective(x, parameters, target, target_tt, summary_function, tolerance)
}
\arguments{
\item{x}{Input parameter to be varied. For malariasimulation this is EIR.}

\item{parameters}{Other malariasimulation model parameters}

\item{target}{Values of target variable to calibrate to.}

\item{target_tt}{Timesteps of target.}

\item{summary_function}{A function that produces a vector of the target variable.}

\item{tolerance}{The average tolerance, in target variable units, with which to match target.}
}
\value{
Difference between output and target.
}
\description{
Quantifies distance to target
}
