% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{Objective function}
\usage{
objective(x, parameters, summary_function)
}
\arguments{
\item{x}{Input parameter to be varied. For malariasimulation this is EIR.}

\item{parameters}{Other malariasimulation model parameters}

\item{summary_function}{A function that take the raw model output as an argument and
produces a vector of the target variable.}
}
\value{
Difference between output and target.
}
\description{
Quantifies distance to target
}
