% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{proposal}
\alias{proposal}
\title{Propose new EIR, moving on log scale}
\usage{
proposal(a, b)
}
\arguments{
\item{a}{lower EIR}

\item{b}{upper EIR}
}
\value{
EIR
}
\description{
Propose new EIR, moving on log scale
}
