% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcali.R
\name{run_simulations}
\alias{run_simulations}
\title{Run simulations from malariasimuation}
\usage{
run_simulations(
  parameters,
  target_tt,
  interval,
  nsims,
  par_name = "init_EIR",
  summary_name = "prev",
  ncores,
  summary_function,
  change_function,
  trans.par,
  inv.trans.par,
  ...
)
}
\arguments{
\item{parameters}{Other malariasimulation model parameters}

\item{target_tt}{Timesteps of target.}

\item{ncores}{Number of simulations to run}

\item{summary_function}{A function that produces a vector of the target variable.}

\item{target}{Values of target variable to calibrate to.}

\item{test_EIRs}{A vector of EIR values to use in the simulations}
}
\value{
A data.table of simulation results
}
\description{
Run simulations from malariasimuation
}
