% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcali.R
\name{score_calibration}
\alias{score_calibration}
\title{Simulate from parameters using calibrated value and calculate continuous ranked probability score}
\usage{
score_calibration(parameters, target, target_tt, nsims = 100, summary_function)
}
\arguments{
\item{parameters}{Other malariasimulation model parameters}

\item{target}{Values of target variable to calibrate to.}

\item{target_tt}{Timesteps of target.}

\item{nsims}{Number of simulations to run}

\item{summary_function}{A function that produces a vector of the target variable.}

\item{ncores}{Number of cpu cores to perform simulations on in parallel}
}
\value{

}
\description{
Simulate from parameters using calibrated value and calculate continuous ranked probability score
}
